requirements.txt
# Usage: pip install -r requirements.txt
opencv-contrib-python
numpy
Pillow
imutils
streamlit

